Kill warning

+++ avs_core/core/parser/expression.h
@@ -108,7 +108,7 @@ class ExpConstant : public Expression
 {
 public:
 #ifndef OLD_ARRAYS
-  ExpConstant(std::vector<AVSValue>* v) : val(v->data(), v->size()) {} // array of AVSValue*
+  ExpConstant(std::vector<AVSValue>* v) : val(v->data(), (int)(v->size())) {} // array of AVSValue*
 #endif
   ExpConstant(AVSValue v) : val(v) {}
   ExpConstant(int i) : val(i) {}
